Move project to module imports. Minor changes.
