+++ 3-bull-jobs/finished-app/lib/movies/movieHelper.js
@@ -1,9 +1,10 @@
-const _ = require('lodash');
+import _ from 'lodash';
 
-const DataHelper = require('../../data/dataHelper');
-const MovieAPIService = require('./movieApiService');
+import { DataHelper } from '../../data/dataHelper.js';
+import { MovieAPIService } from './movieApiService.js';
+import { GenreFetchingQueue } from '../../queues/index.js';
 
-class MovieHelper {
+export class MovieHelper {
   constructor() {
     this.apiService = new MovieAPIService();
     this.dataHelper = new DataHelper();
@@ -36,37 +37,6 @@ class MovieHelper {
       throw new Error('Need at least a genre to filter');
     }
 
-    await this.dataHelper.saveStatus('Loading');
-
-    const validGenres = await this.apiService.getGenres();
-
-    const mappedGenres = new Map(validGenres.map((genre) => [genre.name.toLowerCase(), genre.id]));
-
-    const selectedGenres = [];
-    for (const genre of genres) {
-      const genreId = mappedGenres.get(genre.toLowerCase());
-
-      if (!_.isNil(genreId)) {
-        selectedGenres.push(genreId);
-      }
-    }
-
-    if (_.isEmpty(selectedGenres)) {
-      throw new Error(`Not a valid Genre. Please choose any of the following: ${[...mappedGenres.keys()].join(', ')}`);
-    }
-
-    const movies = await this.apiService.getMoviesByGenres(selectedGenres);
-
-    const sortedMovies = this._sortMoviesByGenre(movies, selectedGenres);
-    await this.dataHelper.saveMovies(sortedMovies);
-    await this.dataHelper.saveStatus('Loaded');
-  }
-
-  _sortMoviesByGenre(movies, genres) {
-    return _.sortBy(movies, (movie) => {
-      return movie.genre_ids.length;
-    });
+    await GenreFetchingQueue.enqueue({ genres });
   }
 }
-
-module.exports = MovieHelper;
